+++ fmriprep/cli/run.py
@@ -65,12 +65,20 @@ def get_parser():
 
     g_conf = parser.add_argument_group('Workflow configuration')
     g_conf.add_argument(
-        '--ignore', required=False, action='store', choices=['fieldmaps',
-                                                             'slicetiming'], nargs="+", default=[],
-        help='ignore selected aspects of the input dataset to disable related'
+        '--ignore', required=False, action='store', nargs="+", default=[],
+        choices=['fieldmaps', 'slicetiming'],
+        help='ignore selected aspects of the input dataset to disable corresponding '
              'parts of the workflow')
-    g_conf.add_argument('--skip-native', action='store_true', default=False,
-                        help="don't output timeseries in native space")
+    g_conf.add_argument('--bold2t1w-dof', action='store', default=9, choices=[6, 9, 12], type=int,
+                        help='Degrees of freedom when registering BOLD to T1w images. '
+                             '9 (rotation, translation, and scaling) is used by '
+                             'default to compensate for field inhomogeneities.')
+    g_conf.add_argument(
+        '--output-space', required=False, action='store',
+        choices=['T1w', 'MNI152NLin2009cAsym',
+                 'fsnative', 'fsaverage', 'fsaverage6', 'fsaverage5'],
+        nargs='+', default=['MNI152NLin2009cAsym', 'fsaverage5'],
+        help='volume and surface spaces to resample functional series into')
 
     #  ANTs options
     g_ants = parser.add_argument_group('Specific options for ANTs registrations')
@@ -81,6 +89,13 @@ def get_parser():
     g_ants.add_argument('--no-skull-strip-ants', dest="skull_strip_ants", action='store_false',
                         help="don't use ANTs-based skull-stripping (use  AFNI instead, fast)")
     g_ants.set_defaults(skull_strip_ants=True)
+
+    # Fieldmap options
+    g_fmap = parser.add_argument_group('Specific options for handling fieldmaps')
+    g_fmap.add_argument('--fmap-bspline', action='store_true', default=False,
+                        help='fit a B-Spline field using least-squares (experimental)')
+    g_fmap.add_argument('--fmap-no-demean', action='store_false', default=True,
+                        help='do not remove median (within mask) from fieldmap')
 
     # FreeSurfer options
     g_fs = parser.add_argument_group('Specific options for FreeSurfer preprocessing')
@@ -127,10 +142,13 @@ def create_workflow(opts):
         'output_dir': op.abspath(opts.output_dir),
         'work_dir': op.abspath(opts.work_dir),
         'ignore': opts.ignore,
-        'skip_native': opts.skip_native,
+        'output_spaces': opts.output_space,
         'freesurfer': opts.freesurfer,
         'hires': opts.hires,
         'reportlets_dir': op.join(op.abspath(opts.work_dir), 'reportlets'),
+        'fmap_bspline': opts.fmap_bspline,
+        'fmap_demean': opts.fmap_no_demean,
+        'bold2t1w_dof': opts.bold2t1w_dof,
     }
 
     # set up logger