+++ CHANGES.rst
@@ -1,9 +1,14 @@
-Next release
-============
+0.4.0 (20th of April 2017)
+==========================
 
+* [ENH] Added an option to choose the degrees of freedom used when doing BOLD to T1w coregistration (``--bold2t1w_dof``). Set default to 9 to account for field inhomogeneities and coils heating up (#448)
+* [ENH] Added support for phase difference and GE style fieldmaps (#448)
 * [ENH] Generate GrayWhite, Pial, MidThickness and inflated surfaces (#398)
 * [ENH] Memory and performance improvements for calculating the EPI reference (#436)
 * [ENH] Sample functional series to subject and ``fsaverage`` surfaces (#391)
+* [ENH] Output spaces for functional data may be selected with ``--output-space`` option (#447)
+* [DEP] ``--skip-native`` functionality replaced by ``--output-space`` (#447)
+* [ENH] ``fmriprep-docker`` wrapper script simplifies running in a Docker environment (#317)
 
 0.3.2 (7th of April 2017)
 =========================